[Multipart] Add a frames -> raw parts parsing sequence
