+++ docker/docker-compose.2204.510.yaml
@@ -12,7 +12,9 @@ services:
     environment:
       - WARN_AS_ERROR_ARG=-Xswiftc -warnings-as-errors
       - IMPORT_CHECK_ARG=--explicit-target-dependency-import-check error
-      - STRICT_CONCURRENCY_ARG=-Xswiftc -strict-concurrency=complete
+      # Disabled strict concurrency checking as currently it's not possible to iterate an async sequence
+      # from inside an actor without warnings.
+      # - STRICT_CONCURRENCY_ARG=-Xswiftc -strict-concurrency=complete
 
   shell:
     image: *image
